tests/wpt/mozilla/tests/mozilla/bluetooth/requestDevice/adapter-not-present.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';
test(function() {
    window.testRunner.setBluetoothMockDataSet('NotPresentAdapter');
    assert_throws(new TypeError(), function () {
        window.navigator.bluetooth.requestDevice({filters: [{services: ['generic_access']}]});
    });
}, 'Reject with TypeError if the adapter is not present.');
</script>
